-----Roles and Validations: --Edit controller/ArticleController.java --Edit entity/Role.java --Edit entity/User.java --Edit resources/templates/article/details.html --Add resources/templates/article/delete.html --Add resources/templates/article/edit.html

+++ src/main/java/martinBlog/controller/ArticleController.java
@@ -7,6 +7,7 @@ import martinBlog.repository.ArticleRepository;
 import martinBlog.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -61,6 +62,15 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        if(!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)) {
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication().getPrincipal();
+
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+
+            model.addAttribute("user", entityUser);
+        }
+
         Article article = this.articleRepository.findOne(id);
 
         model.addAttribute("article", article);
@@ -79,6 +89,10 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if(!isUserAuthorOrAdmin(article)){
+            return  "redirect:/article/" + id;
+        }
+
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
 
@@ -93,6 +107,10 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if(!isUserAuthorOrAdmin(article)){
+            return  "redirect:/article/" + id;
+        }
 
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
@@ -111,6 +129,10 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if(!isUserAuthorOrAdmin(article)){
+            return  "redirect:/article/" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view","article/delete");
 
@@ -126,8 +148,21 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if(!isUserAuthorOrAdmin(article)){
+            return  "redirect:/article/" + id;
+        }
+
         this.articleRepository.delete(article);
 
         return "redirect:/";
+    }
+
+    private boolean isUserAuthorOrAdmin(Article article) {
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
     }
 }
